Separate ProductPhoto and ProfileAvatar image url builders
